Per reviewer's request, moved copy-pasted ImageSchema file from pyspark.ml.image to sparkdl.image to avoid patchin the __path__ variable of the pyspark.ml.image module
