+++ python/sparkdl/__init__.py
@@ -12,18 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-
-# hack to import copy-pasted image schema (to be removed in Spark2.3)
-# TODO remove in Spark2.3
-import os
-import pyspark.ml
-dir_path = os.path.dirname(os.path.realpath(__file__))
-parentdir = os.path.dirname(dir_path)
-pyspark.ml.__path__.append(os.path.join(parentdir, "pyspark", "ml"))
-
-from pyspark.ml.image import ImageSchema
-
 from .graph.input import TFInputGraph
 from .transformers.keras_image import KerasImageFileTransformer
 from .transformers.named_image import DeepImagePredictor, DeepImageFeaturizer
