+++ python/sparkdl/image/image.py
